requirements/requirements.txt
pytest==5.1.2
pre-commit==1.18.2
PyJWT==1.7.1
isort==4.3.21
bcrypt==3.1.7
Faker==2.0.2
pycrypto==2.6.1
requests==2.21.0
pytest-dependency==0.4.0
pytest-html==1.22.0
simple-settings==0.18.0
allure-pytest==2.8.5
pytest-rerunfailures==7.0
elastic-apm==5.2.2
elasticsearch==6.1.1
SQLAlchemy==1.3.18
requests==2.21.0
GeoAlchemy2==0.8.4
Jinja2==2.10
jinjasql==0.1.8
MarkupSafe==1.1.1